+++ CHANGELOG.md
@@ -30,7 +30,17 @@ The category for changes related to documentation, testing and tooling. Also, fo
 
 ## main
 
+## Breaking changes
+
+* Break `TabNav` into `TabNav` for page navigation and `TabPanel` for panel navigation.
+
+    *Kate Higa
+
 ### Misc
+
+* Extract example tag parsing into helper.
+
+    *Kate Higa*
 
 * Expose custom cops and default config for erblint.